Limit the scope of the nearest-loaded-tab search to the current tab group.
When using tab groups, one ordinarily wants to treat each tab group as
its own mostly-separate entity, with no real connection to the other
tab groups in the same window. As such, selecting a tab from a
different group as the "nearest loaded tab" is unexpected, and
generally undesirable.

To avoid that, limit the scope of findClosestLoadedTab() to only the
current tab group, as embodied by the current tabbrowser.visibleTabs
array.

The previous findClosestLoadedTab() searched tabbrowser.mTabs in both
directions from a starting index of the current tab's _tPos variable.
While visibleTabs itself provides a suitable substitute for mTabs,
there is no available equivalent to _tPos - nor can there be, since
the contents of visibleTabs are apparently replaced wholesale every
time the user switches to a different tab group.

To obtain a viable starting index, we therefore scan visibleTabs for
the current tab.

We also check any candidate tab from the "shortcut" or "fallback" code
to make sure it's in visibleTabs. A side effect of this is that we are
no longer guaranteed to be able to find a suitable tab, unless it is
safe to assume that (visibleTabs.length > 1) means that at least one of
the current tab's previousSibling and nextSibling are in visibleTabs.

If we do end up with no suitable tab, we return null, just as if there
had been only one visible tab to begin with. This seems to produce
the intuitively correct behavior in all cases where it actually
happens.