Use aTab.nextSibling rather than aTab as the starting index for findClosestLoadedTab in the case when the tab was closed rather than unloaded.
When searching for the closest tab to a tab which is being closed, the
current tab is removed from the visibleTabs list before the BarTab
onTabClose hook ever gets called, so scanning the array never finds it
and we end up with the wrong index. As a result, in such cases we
would always end up selecting the last (or perhaps last loaded) tab in
the array, even when there were suitable ones closer to the current
tab.

To avoid this, do two things. One, use aTab.nextSibling as the tabIndex
if it's found before aTab is, and test appropriately to make sure we
don't accidentally re-select the current tab as a result. Two, make
sure that the tabIndex is always strictly less than the array length,
so that we can fall back properly in the event that no loaded tab is
found. (I believe this latter is actually a bugfix regardless.)