Prefer the first tab to the left over the second tab to the right when choosing the nearest tab in the case where the current tab was closed.
Since aTab has already been removed from visibleTabs by the time we call
findClosestLoadedTab in the case of onTabClose, and since we therefore
start with an offset of 0 instead of 1, we can end up checking two tabs to
the right before we check one tab to the left. The sequence is:

  scan for tabIndex, and find aTab.nextSibling first (since aTab is not in visibleTabs anymore)
  check tabIndex + 0 (aTab.nextSibling), which is unloaded
  check tabIndex - 0 (aTab.nextSibling), which is unloaded
  check tabIndex + 1 (aTab.nextSibling.nextSibling), which is loaded
  return

And we never checked tabIndex - 1 (aTab.previousSibling), even though it
would be the better choice if it were loaded (since the effective new offset
of the closed aTab is actually tabIndex - 0.5).

To avoid this, in the case of tabIndex+0, we preemptively increment our offset
from 0 to 1, check tabIndex-1, then re-decrement the offset again. I've worked
a number of examples, and they all seem to produce the correct result, both
where the previous code does and where it does not.