Drop Seamonkey support for now.
Seamonkey doesn't implement Panorama, so the entire visibleTabs approach
won't work. To support both Seamonkey and Firefox, two parallel sets of code
would be necessary, including probably two options dialogs.

While this might be doable (especially if we also make the visibleTabs
approach a configurable option even in Firefox), it might also be better to
simply have Seamonkey support be a separate add-on.